requirements.txt
openai==1.30.3
python-dotenv==0.19.2
pandas==2.0.3
streamlit==1.35.0
google-generativeai==0.7.2
weave==0.50.10
